Time: 256 ms (57.14%) | Memory: 0B (100.00%) - LeetSync

+++ 2315-the-change-in-global-rankings/the-change-in-global-rankings.sql
@@ -1,7 +1,6 @@
 SELECT 
     a.team_id, 
     a.name,
-    (ROW_NUMBER() OVER(ORDER BY points DESC, name ASC))::BIGINT - 
-    (ROW_NUMBER() OVER(ORDER BY (points + points_change) DESC, name ASC))::BIGINT AS rank_diff
+    (ROW_NUMBER() OVER(ORDER BY points DESC, name ASC))::INT - (ROW_NUMBER() OVER(ORDER BY (points + points_change) DESC, name ASC))::INT AS rank_diff
 FROM TeamPoints a
 JOIN PointsChange b ON a.team_id = b.team_id;
